+++ dakaraphi.devtools.tracing/src/test/java/dakaraphi/devtools/TestApplication.java
@@ -1,6 +1,8 @@
 // Note this must not exist in 'dakaraphi.devtools.tracing' as that package is excluded from stack trace info in ApplicationHooks
 package dakaraphi.devtools;
 
+import org.junit.Test;
+
 public class TestApplication {
     public static void main(String[] args) throws Exception {
         System.out.println("Test application");
@@ -8,6 +10,9 @@ public class TestApplication {
             method2();
         }
         method3("testing");
+
+        new SimpleObject();
+        new PlainObject();
 
         Thread background = new Thread(()-> backgroundTask());
         background.setDaemon(true);
@@ -32,4 +37,10 @@ public class TestApplication {
 
         }
     }
+    public static class SimpleObject {
+        public SimpleObject() {
+            System.out.println("invoked constructor");
+        }
+    }
 }
+